ai-service/requirements.txt
fastapi
uvicorn
python-multipart

pypdf
numpy
pandas
xgboost
sentence-transformers

google-genai
python-dotenv

